requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
flask==3.0.3
Flask-Cors==4.0.1
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.65.4
gunicorn==23.0.0
idna==3.7
importlib-metadata==8.2.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.24.4
packaging==24.1
pandas==2.0.3
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rsa==4.9
six==1.16.0
soupsieve==2.5
tzdata==2024.1
urllib3==2.2.2
werkzeug==3.0.3
zipp==3.20.0
